+++ app/Mail/FormalNotification.php
@@ -3,16 +3,14 @@
 namespace App\Mail;
 
 use App\DTO\FormalNotificationData;
-use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Mail\Mailable;
 use Illuminate\Mail\Mailables\Content;
 use Illuminate\Mail\Mailables\Envelope;
 use Illuminate\Queue\SerializesModels;
 
-class FormalNotification extends Mailable implements ShouldQueue
+class FormalNotification extends Mailable
 {
-    use Queueable, SerializesModels;
+    use SerializesModels;
 
     public function __construct(public FormalNotificationData $notification) {}